Clustering checks, distance parameter, snp subsample
Added a clustering check function that will de-cluster incorrect clusters based on the haplotypes before and after the recombination breakpoint. Added a way to compute the distance parameter for the hierarchical clustering method using the intervals calculated from the posterior probabilities. Changed the sub sampling method to sub sample based on number of SNPs per genome and not by coverage. Also changed the data structure of the polarized array so it occupies half as much disk space.