Added std.out logs, missing data filter
Added lines writing out to std.out so that you can examine logs of the run. Changed the way missing data is handled so that np arrays will remain the same length, but missing data will be replaced w/ NaNs. Then added a method to detect and delete the NaN missing data for downstream analysis.